requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1